+++ src/main/java/com/goodee/corpdesk/approval/repository/ApprovalFormRepository.java
@@ -3,6 +3,6 @@ package com.goodee.corpdesk.approval.repository;
 import com.goodee.corpdesk.approval.entity.ApprovalForm;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApprovalFormRepository extends JpaRepository<ApprovalForm, Long> {
+public interface ApprovalFormRepository extends JpaRepository<ApprovalForm, Integer> {
 
 }
